+++ packages/typespec-azure-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @azure-tools/typespec-azure-core
 
+## 0.48.0
+
+### Bug Fixes
+
+- [#1545](https://github.com/Azure/typespec-azure/pull/1545) Add docs for linting rules and stop excluding non `Azure.` namespace
+
+### Bump dependencies
+
+- [#1663](https://github.com/Azure/typespec-azure/pull/1663) Upgrade dependencies
+
+
 ## 0.47.0
 
 ### Bug Fixes